requirements.txt
streamlit
pypdf
langchain
langchain_community
langchain-google-genai
tiktoken
sentence-transformers
python-dotenv
